requirements.txt
# Essential dependencies only - no version conflicts
clickhouse-connect
peewee
pandas
numpy
SQLAlchemy
confluent-kafka
kafka-python
python-dotenv
requests
PyYAML
pendulum
croniter
structlog
yfinance
apache-airflow-providers-apache-kafka
streamlit
airflow-clickhouse-plugin